delete : user service interface
