--- src/main/java/com/flab/jbly/presentation/user/response/DeploySuccessResponse.java
@@ -1,17 +1,0 @@
-package com.flab.jbly.presentation.user.response;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class DeploySuccessResponse {
-
-    String text;
-
-    @Builder
-    public DeploySuccessResponse(String text) {
-        this.text = text;
-    }
-}
